requirements.txt
pandas
numpy
python-dotenv
delta_rest_client
requests
flask
flask-cors
flask-socketio
python-socketio
eventlet
feedparser
bs4